improve: Clean up dashboard and move notifications to settings
- Remove test alert and voicemail check buttons from dashboard
- Move notification enable functionality to settings page
- Remove email monitoring section from settings (auto-refresh active)
- Simplify dashboard to focus on alert display
